chatbot/requirements.txt
# Core Web Framework
flask>=2.3.3
flask-cors>=4.0.0

# AI & LLM
openai>=1.35.0

# Machine Learning & Embeddings
sentence-transformers>=2.2.2
faiss-cpu>=1.8.0
langchain-community>=0.2.0
transformers>=4.33.0
torch>=2.2.0

# Data Processing
numpy>=1.24.3
pandas>=2.0.0
scikit-learn>=1.3.0

# Image Processing
Pillow>=10.0.0

# Database
mysql-connector-python>=8.1.0

# Web Scraping & HTTP Requests
requests>=2.31.0
beautifulsoup4>=4.12.2

# Utilities
python-dotenv>=1.0.0
PyJWT>=2.8.0

# Encryption & Security (cho booking links)
cryptography>=41.0.0
pycryptodome>=3.18.0

# Caching (optional - for performance)
redis>=5.0.0

# Date/Time handling
python-dateutil>=2.8.2
